fix: check for number of disks
